requirements.txt
requests>=2.31.0
urllib3>=2.0.0
playwright>=1.40.0
fake-useragent>=1.4.0
pandas>=1.5.0
psycopg2-binary>=2.9.0
